// src/types/types.ts
export interface Film {
  title: string;
  year: number;
  month: number;
  role: string;
  type: 'Long Métrage' | 'Série TV' | 'Publicité' | 'Court Métrage' | 'Documentaire';
  impactScore: number; // 1 to 5
  boxOffice?: number; // entrées en salle
  audience?: number; // spectateurs
}

export enum Availability {
  AVAILABLE = 'Disponible',
  UNAVAILABLE = 'Indisponible',
  SOON = 'Bientôt disponible',
}

export interface Photo {
  id: number;
  url: string;
}

export interface Payment {
  id: number;
  year: number;
  amount: number;
  status: 'Payée' | 'Impayée';
  date: string; // ISO String
}

export interface Technician {
  id: number;
  name: string;
  specialty: string;
  avatarUrl: string;
  availability: Availability;
  bio: string;
  filmography: Film[];
  email: string;
  phone: string;
  skills: string[];
}

export type NotificationPreference = 'email' | 'sms' | 'none';

export type UserRole = 
  'Directeur Exécutif' | 
  'Président du CA' | 
  'Trésorière' | 
  'Membre' |
  'Secrétaire Général' |
  'Secrétaire Général Adjoint' |
  'Secrétaire à la communication' |
  'Secrétaire à la communication Adjoint' |
  'Trésorier Adjoint';

export type MemberStatus = 'Membre Actif' | 'Membre d\'Honneur' | 'Membre Bienfaiteur' | 'Sanctionné';


export interface Member extends Technician {
  membershipPaid: boolean;
  nextMeeting: string;
  notificationPreference: NotificationPreference;
  gallery: Photo[];
  paymentHistory: Payment[];
  role?: UserRole;
  status: MemberStatus;
  twoFactorEnabled: boolean;
  passwordHash: string; // NOUVEAU: pour simuler un mot de passe haché
}
// ...
export interface Notification {
  id: number;
  userId: number;
  message: string;
  date: string;
  read: boolean;
  type: 'info' | 'warning' | 'success';
  link?: string;
}